Regenerate Android.bp using go2bp, fully switch to new protobuf package
Android is upgrading directly from the deprecated Go protobuf package to the
new package, as we don't have very much Go/protobuf code. The newer
versions of the deprecated package are just thin wrappers around the new
package.

I used the go2bp to automate the handling of the Android.bp Go modules,
and it should make future upgrades simpler.

The *.pb.go files were hand edits, though they appear to be the only
functional changes as regenerating with a newer protoc-gen-go. The
existing files do not match using Android's protoc directly, as the
comments are passed through on the Android version. It's possible
there's a version/flag mismatch, or the bazel rules are stripping
comments from the proto before passing it to protoc.

Change-Id: Ifb92f17aea0c2c7d4dd9217555a393a500e62a9a